Move logger.ts inside lib
